Improve day 15 solution
Compute contiguous exclusion row segments instead of overlapping ones.
This bypasses the need to sort the exclusion segments every time, and
also creates shorter vectors of segments: basically only one segment for
almost all rows in y=0..4M, except for the line that has the distress
beacon, which has two segments.

This decreases the runtime of day 15 from ~350ms to ~200ms. Not too
impressive, but it's nice :)