Unify Dijkstra's algorithm functions
Remove the function that calculates all distances from a starting node
in favor of the one that computes a single shortest distance from a
starting point to and end goal. Adapt day 12 and 16 solutions to use the
latter.

The results of these solutions is still the same, and the runtime
doesn't seem to differ in any significant way. I also added some sample
tests for those two solutions.